+++ src/accessDeep.ts
